Use ImageFile directly with KWCOCO ETL
